requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pycups==2.0.4
pyserial==3.5
python-barcode==0.15.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.2
qrcode==8.2
setuptools==80.9.0
six==1.17.0
websockets==15.0.1
Werkzeug==3.1.3
numpy==2.2.6
pygame==2.6.1
ur_rtde==1.6.1
